requirements.txt
# Danbooru-related
pybooru

# General
tqdm
faiss-cpu
numpy
requests
transformers
sentence-transformers
diffusers
torch
triton
ollama
accelerate
PyYAML
pydantic-settings

# LoRA training
peft
bitsandbytes
